+++ tests/runtime/engine/runner.py
@@ -162,7 +162,6 @@ class Runner(object):
         bpffeature["aot"] = cmake_vars.LIBBCC_BPF_CONTAINS_RUNTIME
         bpffeature["skboutput"] = output.find("skboutput: yes") != -1
         bpffeature["get_tai_ns"] = output.find("get_ktime_ns: yes") != -1
-        bpffeature["get_func_ip"] = output.find("get_func_ip: yes") != -1
         bpffeature["jiffies64"] = output.find("jiffies64: yes") != -1
         bpffeature["lookup_percpu_elem"] = output.find("lookup_percpu_elem: yes") != -1
         return bpffeature